Adjust to modified spindle related OM structure
Move ObjectModel related stuff into sub-dir and split out
